Correct batch size format

+++ train.py
@@ -120,8 +120,8 @@ val_dataset = TrainDataset('data/vqa_v2','val','VQAv2')
 train_targ_dataset = TestDataset('data/test/images', 'data/test/train_questions.csv')
 test_targ_dataset = TestDataset('data/test/images', 'data/test/test_questions.csv')
 
-train_loader = DataLoader(train_dataset, num_workers=num_workers, batch_size=batch_size*0.75, shuffle=False)
-train_targ_loader = DataLoader(train_targ_dataset, num_workers=num_workers, batch_size=batch_size*0.25, shuffle=False)
+train_loader = DataLoader(train_dataset, num_workers=num_workers, batch_size=int(batch_size*0.75), shuffle=False)
+train_targ_loader = DataLoader(train_targ_dataset, num_workers=num_workers, batch_size=int(batch_size*0.25), shuffle=False)
 val_loader = DataLoader(val_dataset, num_workers=num_workers, batch_size=batch_size, shuffle=False)
 # cross_loader = DataLoader(cross_dataset, num_workers=num_workers, batch_size=batch_size, shuffle=False)
